+++ src/metric/src/Adapter/NoOp/MetricFactory.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
+
+namespace Hyperf\Metric\Adapter\NoOp;
+
+use Hyperf\Contract\ConfigInterface;
+use Hyperf\Metric\Contract\CounterInterface;
+use Hyperf\Metric\Contract\GaugeInterface;
+use Hyperf\Metric\Contract\HistogramInterface;
+use Hyperf\Metric\Contract\MetricFactoryInterface;
+
+class MetricFactory implements MetricFactoryInterface
+{
+    /**
+     * @var ConfigInterface
+     */
+    private $config;
+
+    public function __construct(ConfigInterface $config)
+    {
+        $this->config = $config;
+    }
+
+    public function makeCounter(string $name, ?array $labelNames = []): CounterInterface
+    {
+        return new Counter();
+    }
+
+    public function makeGauge(string $name, ?array $labelNames = []): GaugeInterface
+    {
+        return new Gauge();
+    }
+
+    public function makeHistogram(string $name, ?array $labelNames = []): HistogramInterface
+    {
+        return new Histogram();
+    }
+
+    public function handle(): void
+    {
+    }
+}